Prototipo_XGBoost_SHAP_Cloud_IA/requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
pyyaml
joblib
matplotlib
streamlit
qdrant-client
huggingface-hub
requests
python-dotenv
